allocation record card
