+++ app/components/allocation_record_card.html.erb
